fixup: address Kurt's comments
